career save

+++ Hammer_Time/Assets/Scripts/Tourny/Team.cs
@@ -15,5 +15,5 @@ public class Team
     public int strength;
     public int id;
 
-    public int earnings;
+    public float earnings;
 }
